fix: remove auto import & setting multiprocessing context
