+++ kedro/framework/session/store.py
@@ -1,10 +1,6 @@
 """This module implements a dict-like store object used to persist Kedro sessions."""
-import dbm
 import logging
-import shelve
 from collections import UserDict
-from multiprocessing import Lock
-from pathlib import Path
 from typing import Any, Dict
 
 
@@ -41,35 +37,3 @@ class BaseSessionStore(UserDict):
             "'save()' not implemented for '%s'. Skipping the step.",
             self.__class__.__name__,
         )
-
-
-class ShelveStore(BaseSessionStore):
-    """Stores the session data on disk using `shelve` package."""
-
-    _lock = Lock()
-
-    @property
-    def _location(self) -> Path:
-        return Path(self._path).expanduser().resolve() / self._session_id / "store"
-
-    def read(self) -> Dict[str, Any]:
-        """Read the data from disk using `shelve` package."""
-        data = {}  # type: Dict[str, Any]
-        try:
-            with shelve.open(str(self._location), flag="r") as _sh:  # nosec
-                data = dict(_sh)
-        except dbm.error:
-            pass
-        return data
-
-    def save(self) -> None:
-        """Save the data on disk using `shelve` package."""
-        location = self._location
-        location.parent.mkdir(parents=True, exist_ok=True)
-
-        with self._lock, shelve.open(str(location)) as _sh:  # nosec
-            keys_to_del = _sh.keys() - self.data.keys()
-            for key in keys_to_del:
-                del _sh[key]
-
-            _sh.update(self.data)
